SamIntake/requirements.txt
click==8.0.0
cycler==0.10.0
Flask==2.0.0
Flask-JSON==0.3.4
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python-headless==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
ulid-py==1.1.0
Werkzeug==2.0.1
